move AjnaPosition class out of the types
