+++ packages/oasis-actions/src/helpers/ajna/index.ts
@@ -1,0 +1,1 @@
+export { AjnaPosition } from './AjnaPosition'
