+++ packages/oasis-actions/src/strategies/ajna/depositBorrow.ts
@@ -1,7 +1,7 @@
 import * as ethers from 'ethers'
 
 import ajnaProxyActionsAbi from '../../abi/ajna/ajnaProxyActions.json'
-import { AjnaPosition } from '../../types/ajna'
+import { AjnaPosition } from '../../helpers/ajna/AjnaPosition'
 import { Strategy } from '../../types/common'
 import { Dependencies, OpenArgs } from './open'
